move gamebadge into badge
